Rename qabel-register into qabel-index
