+++ index_service/templates/verification/email.tpl
