+++ index_service/templates/verification/sms.tpl
