+++ pytest.ini
@@ -1,9 +1,9 @@
 [pytest]
-DJANGO_SETTINGS_MODULE=qabel_register.settings
+DJANGO_SETTINGS_MODULE=qabel_index.settings
 testpaths=tests
 python_files=*.py
 python_classes=*Test
 
 flake8-max-line-length=140
 flake8-ignore=
-    register_service/migrations/* ALL
+    index_service/migrations/* ALL
