+++ qabel_index/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for qabel_register project.
+Django settings for qabel_index project.
 
 Generated by 'django-admin startproject' using Django 1.8.4.
 
@@ -41,7 +41,7 @@ INSTALLED_APPS = (
     'mail_templated',
     'rest_framework',
     'sendsms',
-    'register_service',
+    'index_service',
 )
 
 MIDDLEWARE_CLASSES = (
@@ -55,7 +55,7 @@ MIDDLEWARE_CLASSES = (
     'django.middleware.security.SecurityMiddleware',
 )
 
-ROOT_URLCONF = 'qabel_register.urls'
+ROOT_URLCONF = 'qabel_index.urls'
 
 TEMPLATES = [
     {
@@ -73,7 +73,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'qabel_register.wsgi.application'
+WSGI_APPLICATION = 'qabel_index.wsgi.application'
 
 
 # Database
@@ -93,7 +93,7 @@ else:
     DATABASES = {
         'default': {
             'ENGINE': 'django.db.backends.postgresql_psycopg2',
-            'NAME': 'qabel_register',
+            'NAME': 'qabel_index',
             'USER': 'qabel',
             'PASSWORD': 'qabel_test',
             'HOST': '127.0.0.1',
